requirements.txt
Django==3.0.3
djangorestframework==3.11.0
django-rest-swagger==2.2.0
